rename controll classes names
