+++ src/main/java/com/syst/trades/controller/ClientTradeController.java
@@ -8,7 +8,6 @@ import javax.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,36 +15,30 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import com.syst.trades.model.Sale;
-import com.syst.trades.repository.SaleRepository;
+import com.syst.trades.model.ClientTradeDebit;
+import com.syst.trades.repository.ClientTradeDebitRepository;
 
 @RestController
-@RequestMapping("/sale")
-public class SaleResource {
+@RequestMapping("/client-trade")
+public class ClientTradeController {
 
 	@Autowired
-	private SaleRepository saleRepository;
+	private ClientTradeDebitRepository clientTradeDebitRepository;
 
 	@GetMapping
-	public List<Sale> toList(){
-		return saleRepository.findAll();
-	} // esse cara tem que ser refatorado por trade_id
+	public List<ClientTradeDebit> toList() {
+		return clientTradeDebitRepository.findAll();
+	}
 
 	@PostMapping
 	@ResponseStatus(HttpStatus.CREATED)
-	public void toSave(@RequestBody Sale sale, HttpServletResponse response) {
-		Sale saleSaved = saleRepository.save(sale);
+	public void toSave(@RequestBody ClientTradeDebit clientTradeDebit, HttpServletResponse response) {
+		ClientTradeDebit clientTradeDebitSaved = clientTradeDebitRepository.save(clientTradeDebit);
 
 		URI uri = ServletUriComponentsBuilder.fromCurrentRequestUri().path("/{id}")
-				.buildAndExpand(saleSaved.getId()).toUri();
+				.buildAndExpand(clientTradeDebitSaved.getId()).toUri();
 
 		response.setHeader("Location", uri.toASCIIString());
-	}
-
-	@GetMapping("/status/{statusId}")
-	@ResponseStatus(HttpStatus.OK)
-	public List<Sale> getSaleByStatusId(@PathVariable("statusId") Long statusId) {
-		return saleRepository.findByStatusId(statusId);
 	}
 
 }
