+++ src/main/java/com/syst/trades/controller/PaymentModeController.java
@@ -9,34 +9,34 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import com.syst.trades.model.HistoryProductQuantity;
-import com.syst.trades.repository.HistoryProductQuantityRepository;
+import com.syst.trades.model.PaymentMode;
+import com.syst.trades.repository.PaymentModeRepository;
 
 @RestController
-@RequestMapping("/history-product-quantity")
-public class HistoryProductQuantityResource {
+@RequestMapping("/payment-mode")
+public class PaymentModeController {
 
 	@Autowired
-	private HistoryProductQuantityRepository historyProductQuantityRepository;
+	private PaymentModeRepository paymentModeRepository;
 
 	@GetMapping
-	public List<HistoryProductQuantity> toList() {
-		return historyProductQuantityRepository.findAll();
+	public List<PaymentMode> toList() {
+		return paymentModeRepository.findAll();
 	}
 
 	@PostMapping
 	@ResponseStatus(HttpStatus.CREATED)
-	public void toSave(HistoryProductQuantity historyProductQuantity, HttpServletResponse response) {
-		HistoryProductQuantity historyProductQuantitySaved =
-				historyProductQuantityRepository.save(historyProductQuantity);
+	public void toSave(@RequestBody PaymentMode paymentMode, HttpServletResponse response) {
+		PaymentMode paymentModeSaved = paymentModeRepository.save(paymentMode);
 
 		URI uri = ServletUriComponentsBuilder.fromCurrentRequestUri().path("/{id}")
-				.buildAndExpand(historyProductQuantitySaved.getId()).toUri();
+				.buildAndExpand(paymentModeSaved.getId()).toUri();
 
 		response.setHeader("Location", uri.toASCIIString());
 	}
